Changing in models, updating firebase, refactoring cast screen and refactoring icons with logo in navbar
